pyqtgraph_serial_viewer/requirements.txt
PyQt5>=5.15
pyqtgraph>=0.13
pyserial>=3.5
numpy>=1.20
